requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
Authlib==1.2.1
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
Babel==2.13.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
beniget==0.4.1
bidict==0.22.1
blinker==1.7.0
boto3==1.28.62
botocore==1.31.62
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
circuitbreaker==2.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==41.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dlib==19.24.2
dnspython==2.4.2
duplicity==0.8.21
email-validator==2.0.0.post2
eventlet==0.33.3
exceptiongroup==1.1.3
fastapi==0.104.1
fasteners==0.14.1
Flask==3.0.0
flask-babel==3.1.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.1.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gpg==1.16.0
greenlet==1.1.2
h11==0.14.0
html5lib==1.1
httpagentparser==1.9.5
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
natsort==8.0.2
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
pgadmin4==7.7
Pillow==9.0.1
ply==3.11
portalocker==2.8.2
powerline-status==2.7
protobuf==4.24.4
psutil==5.9.5
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybreaker==1.0.2
pycairo==1.20.1
pycircuitbreaker==0.5.0
pycparser==2.21
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyotp==2.9.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-engineio==4.7.1
python-socketio==5.9.0
pythran==0.10.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4.2
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
scipy==1.8.0
SecretStorage==3.3.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
speaklater3==1.4
SQLAlchemy==2.0.21
sqlparse==0.4.4
sshtunnel==0.4.0
starlette==0.27.0
sympy==1.9
systemd-python==234
typing_extensions==4.8.0
ua-parser==0.18.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==2.0.7
usb-creator==0.3.7
user-agents==2.2.0
uvicorn==0.24.0.post1
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.0.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
